Added new project; same as old w/o unit tests
Running Mariano's project for some reason was making VS so slow it was
unusable, so I just created a new project and copied and pasted his work
into this one.  I kept getting a linker error every time i tried to make
unit testing work (error showed up when i added a reference to the .dll
project in the test project settings so i just gave up on that).  Just
going to work on this one for now, maybe eventually we can copy and
paste my work into mariano's version or vice versa.