Filled in helper functions
I wasn't sure what to do with the unit test files so I just left them
alone. I simply added the helper functions (addColumns(), addRow(),
etc). I'm a bit lost on what we have accomplished so far (not to be rude
or anything). If someone could fill the group in on exactly which
functions we have completed and not would be nice. Sorry for not working
on it as much this weekend; Did the best I could with the helper
functions